+++ src/api/credential-api.js
@@ -1,17 +1,11 @@
-import { axiosInstance as axios, handleError } from "@/api/axios-instance";
-import qs from "qs";
+import { request, handleError } from "@/api/http";
 
 /**
  * Returns a full list of credentials.
  */
 export function getCredentials({ tags, search } = {}) {
-  return axios
-    .get("/credentials", {
-      params: { tags, search },
-      paramsSerializer: (p) =>
-        qs.stringify(p, { arrayFormat: "repeat", skipNulls: true }),
-    })
-    .then(({ data }) => data.records)
+  return request("/credentials", { params: { tags, search } })
+    .then((data) => data.records)
     .catch((error) => Promise.reject(handleError(error)));
 }
 
@@ -19,19 +13,17 @@ export function getCredentials({ tags, search } = {}) {
  * Returns a single credential.
  */
 export function getCredential(id) {
-  return axios
-    .get(`/credentials/${id}`)
-    .then(({ data }) => data)
-    .catch((error) => Promise.reject(handleError(error)));
+  return request(`/credentials/${id}`).catch((error) =>
+    Promise.reject(handleError(error)),
+  );
 }
 
 /**
  * Updates a credential.
  */
 export function updateCredential(id, options) {
-  return axios
+  return request
     .put(`/credentials/${id}`, options)
-    .then(({ data }) => data)
     .catch((error) => Promise.reject(handleError(error)));
 }
 
@@ -39,9 +31,8 @@ export function updateCredential(id, options) {
  * Create a credential.
  */
 export function createCredential(options) {
-  return axios
+  return request
     .post("/credentials", options)
-    .then(({ data }) => data)
     .catch((error) => Promise.reject(handleError(error)));
 }
 
@@ -49,27 +40,25 @@ export function createCredential(options) {
  * Delete a credential.
  */
 export function deleteCredential(id) {
-  return axios
+  return request
     .delete(`/credentials/${id}`)
     .catch((error) => Promise.reject(handleError(error)));
 }
 
 export function deleteTag(credentialId, tag) {
-  return axios
+  return request
     .delete(`credentials/${credentialId}/tags/${tag}`)
     .catch((error) => Promise.reject(handleError(error)));
 }
 
 export function updateTag(credentialId, tag) {
-  return axios
+  return request
     .put(`credentials/${credentialId}/tags/${tag.id}`, tag)
-    .then(({ data }) => data)
     .catch((error) => Promise.reject(handleError(error)));
 }
 
 export function addTag(credentialId, tag) {
-  return axios
+  return request
     .post(`credentials/${credentialId}/tags`, tag)
-    .then(({ data }) => data)
     .catch((error) => Promise.reject(handleError(error)));
 }