+++ src/api/listener-api.js
@@ -1,22 +1,20 @@
-import { axiosInstance as axios, handleError } from "@/api/axios-instance";
+import { request, handleError } from "@/api/http";
 
 /**
  * Returns a single listener.
  */
 export function getListener(id) {
-  return axios
-    .get(`/listeners/${id}`)
-    .then(({ data }) => data)
-    .catch((error) => Promise.reject(handleError(error)));
+  return request(`/listeners/${id}`).catch((error) =>
+    Promise.reject(handleError(error)),
+  );
 }
 
 /**
  * Returns a full list of listeners.
  */
 export function getListeners() {
-  return axios
-    .get("/listeners")
-    .then(({ data }) => data.records)
+  return request("/listeners")
+    .then((data) => data.records)
     .catch((error) => Promise.reject(handleError(error)));
 }
 
@@ -25,10 +23,9 @@ export function getListeners() {
  * @param {string} type the type of listener
  */
 export function getListenerTemplate(templateId) {
-  return axios
-    .get(`/listener-templates/${templateId}`)
-    .then(({ data }) => data)
-    .catch((error) => Promise.reject(handleError(error)));
+  return request(`/listener-templates/${templateId}`).catch((error) =>
+    Promise.reject(handleError(error)),
+  );
 }
 
 /**
@@ -36,9 +33,8 @@ export function getListenerTemplate(templateId) {
  * @param {Object} options the options needed for creating a listener.
  */
 export function createListener(template, options) {
-  return axios
+  return request
     .post("/listeners", { name: options.Name, template, options })
-    .then(({ data }) => data)
     .catch((error) => Promise.reject(handleError(error)));
 }
 
@@ -48,9 +44,8 @@ export function createListener(template, options) {
  */
 export function updateListener(listener) {
   listener.name = listener.options.Name;
-  return axios
+  return request
     .put(`/listeners/${listener.id}`, listener)
-    .then(({ data }) => data)
     .catch((error) => Promise.reject(handleError(error)));
 }
 
@@ -58,9 +53,8 @@ export function updateListener(listener) {
  * Get the listener templates as an array of strings.
  */
 export function getListenerTemplates() {
-  return axios
-    .get("/listener-templates")
-    .then(({ data }) => data.records)
+  return request("/listener-templates")
+    .then((data) => data.records)
     .catch((error) => Promise.reject(handleError(error)));
 }
 
@@ -69,41 +63,37 @@ export function getListenerTemplates() {
  * @param {string} name name of the listener to kill
  */
 export function killListener(name) {
-  return axios
+  return request
     .delete(`/listeners/${name}`)
     .catch((error) => Promise.reject(handleError(error)));
 }
 
 export function deleteTag(listenerId, tag) {
-  return axios
+  return request
     .delete(`listeners/${listenerId}/tags/${tag}`)
     .catch((error) => Promise.reject(handleError(error)));
 }
 
 export function updateTag(listenerId, tag) {
-  return axios
+  return request
     .put(`listeners/${listenerId}/tags/${tag.id}`, tag)
-    .then(({ data }) => data)
     .catch((error) => Promise.reject(handleError(error)));
 }
 
 export function addTag(listenerId, tag) {
-  return axios
+  return request
     .post(`listeners/${listenerId}/tags`, tag)
-    .then(({ data }) => data)
     .catch((error) => Promise.reject(handleError(error)));
 }
 
 export function getAutorunTasks(listenerId) {
-  return axios
-    .get(`/listeners/${listenerId}/autorun`)
-    .then(({ data }) => data.records)
+  return request(`/listeners/${listenerId}/autorun`)
+    .then((data) => data.records)
     .catch((error) => Promise.reject(handleError(error)));
 }
 
 export function saveAutorunTasks(listenerId, modules) {
-  return axios
+  return request
     .put(`/listeners/${listenerId}/autorun`, { records: modules })
-    .then(({ data }) => data)
     .catch((error) => Promise.reject(handleError(error)));
 }